kafka: rename .hh header file to .h
