+++ src/v/kafka/server/fetch_metadata_cache.h
